Mememaster/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
PyMySQL==0.9.3
pyparsing==2.4.0
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.24.3
Werkzeug==0.15.2
xpinyin==0.5.6
